pyproject.toml
[tool.poetry]
name = "nonebot_plugin_fortune"
version = "0.1.0"
description = "Fortune divination!"
authors = ["KafCoppelia <k740677208@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
ujson = "*"
pillow = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
